fix query in graphQLEndpointRetrieveGraphReq, add service map, additional reorganizing
