+++ src/js/engagement_view/src/components/login/checkCookieWhenRouting.tsx
@@ -1,4 +1,4 @@
-import { RouteState } from '../../../types/CustomTypes';
+import { RouteState } from '../../types/CustomTypes';
 
 
 export const defaultRouteState = (): RouteState => {
